Add UniversalBlobStorageFileAccessor project
includes a universal FileAccessor class that can read from or write to
any file in Azure Blob Storage.

NOTE: Normally launchSettings.json is not included in source control as
it sets the environment to "Development".  However, we're always going
to run this demo app in the Development environment so adding it to
source control is not a problem in this case.